Change structure of cloud in for heroku
