+++ cloud/Procfile
@@ -1,1 +1,4 @@
-web: gunicorn --bind 0.0.0.0:$PORT cloud:cloud
+web: gunicorn app:app --bind 0.0.0.0:8000 --log-file=-
+init: flask db init
+migrate: flask db migrate
+upgrade: flask db upgrade
