+++ cloud/requirements.txt
@@ -1,3 +1,7 @@
-Flask==2.0.2
-Flask-Bootstrap==3.3.7.1
-Flask-SQLAlchemy==2.5.1
+alembic
+Flask
+Flask-Bootstrap
+Flask-SQLAlchemy
+Flask-Migrate
+Flask-Script
+gunicorn
